Rename 'util' to 'common'
